instances grass_fed_butter and organic_butter were NOT merged together
Query: in the Butter class, should the instances 'Grass-fed Butter' and 'Organic Butter' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	As a dietician, my answer is no. Although 'Grass-fed Butter' and 'Organic Butter' may share similarities, they are not semantic or ontological duplicates.

Grass-fed butter refers to the diet of the cows, which are fed mainly on grass, while organic butter refers to the production method, adhering to organic farming practices. Both can be grass-fed and organic, but they are not the same, as grass-fed does not necessarily mean organic, and organic does not necessarily mean grass-fed. They can coexist as distinct types of butter.
meaning that instances grass_fed_butter and organic_butter are different.
Query cache in files:
- /kgfiller/data/cache-783b610fabb2e797a997ffc075214874ec0880e2cf586a79fa8a1ba4cdb14e12.yml