instances yogurt and kefir_yogurt were NOT merged together
Query: in the Yogurt class, should the instances 'Kefir Yogurt' and 'Yogurt' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	As a dietician, my answer is:

No, the instances 'Kefir Yogurt' and 'Yogurt' should not be merged together as semantic and ontologic duplicates. They are two distinct types of fermented dairy products with some similarities but also differences in their production process, taste, and probiotic cultures. Kefir Yogurt is a type of yogurt that incorporates kefir grains, resulting in a unique taste and texture. While they both belong to the broader category of yogurt, they have enough differences to warrant separate instances in a classification system.
meaning that instances yogurt and kefir_yogurt are different.
Query cache in files:
- /kgfiller/data/cache-5175718c5195e01316f6094bb4ce04075ae6589fefaefb56aefb87858805da73.yml