instances swordfish and sashimi_grade_fish were NOT merged together
Query: in the Edible Fish class, should the instances 'Sashimi-grade fish' and 'Swordfish' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No, "Sashimi-grade fish" and "Swordfish" should not be merged together as semantic and ontologic duplicates.

"Sashimi-grade fish" is a quality of fish that is suitable for serving raw, while "Swordfish" is a specific type of fish. Although swordfish can be sashimi-grade, not all sashimi-grade fish are swordfish. They represent different concepts in the Edible Fish class.
meaning that instances swordfish and sashimi_grade_fish are different.
Query cache in files:
- /kgfiller/data/cache-0862de7f7e805da58ab8aa44d527a48eb48f934670acda2e7f2c35cdeeab1a44.yml