instances bell_peppers and red_bell_pepper were NOT merged together
Query: in the Edible Vegetable class, should the instances 'Bell peppers' and 'Red bell pepper' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	As a dietician, my answer is:

Yes, it would be appropriate to merge 'Bell peppers' and 'Red bell pepper' as semantic and ontologic duplicates, as red bell peppers are a variety of bell peppers and share the same nutritional properties. However, it is essential to retain the specific information about the color of the bell peppers in the database to allow users to differentiate between different colors of bell peppers and their unique nutritional profiles.
meaning that instances bell_peppers and red_bell_pepper are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-56de1b9db14c2f00254706614ca9f5bb7ee9de3481aed4d6e9d87d02d39b137d.yml